added demo account functionality

+++ src/app/services/budgets/budgets.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { Budget } from '../../types/budget';
 import { BehaviorSubject, map, Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '../../../environments/environment';
+import { AuthService } from '../auth/auth.service';
 
 @Injectable({
   providedIn: 'root',
@@ -20,6 +21,8 @@ export class BudgetsService {
 
   baseURL = `${environment.apiUrl}/api/v1/budgets`;
 
+  authService = inject(AuthService);
+
   constructor(private http: HttpClient) {
     this.loadBudgets();
   }
@@ -34,23 +37,52 @@ export class BudgetsService {
   }
 
   addBudget(newBudget: Budget) {
-    this.http.post<Budget>(this.baseURL, newBudget).subscribe({
-      error: (e) => console.log('An error has occurred', e),
-      complete: () => {
-        this.loadBudgets();
-        console.log('Budget added successfully');
-      },
-    });
+    if (this.authService.getDemo()) {
+      this.demoAdd(newBudget);
+    } else {
+      this.http.post<Budget>(this.baseURL, newBudget).subscribe({
+        error: (e) => console.log('An error has occurred', e),
+        complete: () => {
+          this.loadBudgets();
+          console.log('Budget added successfully');
+        },
+      });
+    }
+  }
+
+  demoAdd(newBudget: Budget) {
+    this.loading.next(true);
+    this.budgets.next([...this.budgets.getValue(), { ...newBudget, id: 0 }]);
+    this.loading.next(false);
   }
 
   removeBudget(id: number) {
-    let tempBudget = this.budgets.getValue();
-    tempBudget = tempBudget.filter((value) => value.id !== id);
-    this.budgets.next(tempBudget);
-    this.http.delete<any>(this.baseURL + `/${id}`).subscribe({
-      error: (e) => console.log('An error has occurred', e),
-      complete: () => console.log('Budget deleted successfully'),
-    });
+    if (this.authService.getDemo()) {
+      this.demoRemove(id);
+    } else {
+      this.loading.next(true);
+      let tempBudget = this.budgets.getValue();
+      tempBudget = tempBudget.filter((value) => value.id !== id);
+      this.http.delete<any>(this.baseURL + `/${id}`).subscribe({
+        next: () => {
+          console.log('Budget deleted successfully');
+          this.budgets.next(tempBudget);
+          this.loading.next(false);
+        },
+        error: (e) => {
+          console.log('An error has occurred', e), this.loading.next(false);
+        },
+      });
+    }
+  }
+
+  demoRemove(id: number) {
+    this.loading.next(true);
+    let newBudgets: Budget[] = this.budgets
+      .getValue()
+      .filter((budget) => budget.id !== id);
+    this.budgets.next(newBudgets);
+    this.loading.next(false);
   }
 
   getBudget(id: number) {
@@ -64,13 +96,33 @@ export class BudgetsService {
   }
 
   updateBudget(updates: Partial<Budget>, id: number) {
+    if (this.authService.getDemo()) {
+      this.demoUpdate(updates, id);
+    } else {
+      this.loading.next(true);
+      const budgets = this.budgets.getValue();
+      const budgetIndex = budgets.findIndex((budget) => budget.id === id);
+      budgets[budgetIndex] = { ...budgets[budgetIndex], ...updates };
+      this.http.patch<any>(this.baseURL + `/${id}`, updates).subscribe({
+        next: () => {
+          console.log('Budget deleted successfully');
+          this.budgets.next(budgets);
+          this.loading.next(false);
+        },
+        error: (e) => {
+          console.log('An error has occurred', e);
+          this.loading.next(false);
+        },
+      });
+    }
+  }
+
+  demoUpdate(updates: Partial<Budget>, id: number) {
+    this.loading.next(true);
     const budgets = this.budgets.getValue();
     const budgetIndex = budgets.findIndex((budget) => budget.id === id);
     budgets[budgetIndex] = { ...budgets[budgetIndex], ...updates };
     this.budgets.next(budgets);
-    this.http.patch<any>(this.baseURL + `/${id}`, updates).subscribe({
-      error: (e) => console.log('An error has occurred', e),
-      complete: () => console.log('Budget deleted successfully'),
-    });
+    this.loading.next(false);
   }
 }
